增加同步主程序和每日推荐同步功能

+++ GeneralMusicDownloader.py
@@ -22,7 +22,7 @@ def random_int_string(length):
     return '%d' % random.randint(start, end)
 
 
-class GeneralMusicDownloader():
+class GeneralMusicDownloader:
     default_headers = {
         'Host': '122.112.253.137',
         'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:57.0) Gecko/20100101 Firefox/57.0',
@@ -71,14 +71,20 @@ class GeneralMusicDownloader():
         if source not in self.source_list:
             print 'Unsupported source expected the one of that: %s' % self.source_list
             return False
+        query_song_data = self.query_song_data.copy()
         callback_item = 'jQuery%s_%s' % (random_int_string(21), timestamp_ms_string())
-        self.query_song_data['callback'] = callback_item
-        self.query_song_data['_'] = timestamp_ms_string()
-        self.query_song_data['name'] = search_keyword
-        self.query_song_data['source'] = source
+        query_song_data['callback'] = callback_item
+        query_song_data['_'] = timestamp_ms_string()
+        query_song_data['name'] = search_keyword
+        query_song_data['source'] = source
 
-        resp = self.session.get(url=self.api_url, params=self.query_song_data, headers=self.default_headers)
-        songs = json.loads(resp.content[len(callback_item) + 1: -1])
+        resp = self.session.get(url=self.api_url, params=query_song_data, headers=self.default_headers)
+        songs = []
+        try:
+            songs = json.loads(resp.content[len(callback_item) + 1: -1])
+        except:
+            traceback.print_exc()
+            print resp.content
         # 响应示例
         #  [
         # 	"id": "001ZptHS0QElqW",
@@ -94,14 +100,15 @@ class GeneralMusicDownloader():
 
     def get_song_download_info_from_source(self, song_id, source):
         # 修改必要参数
+        general_query_data = self.general_query_data.copy()
         callback_item = 'jQuery%s_%s' % (random_int_string(21), timestamp_ms_string())
-        self.general_query_data['callback'] = callback_item
-        self.general_query_data['_'] = timestamp_ms_string()
-        self.general_query_data['id'] = song_id
-        self.general_query_data['source'] = source
-        self.general_query_data['types'] = 'url'
+        general_query_data['callback'] = callback_item
+        general_query_data['_'] = timestamp_ms_string()
+        general_query_data['id'] = song_id
+        general_query_data['source'] = source
+        general_query_data['types'] = 'url'
         # 获取下载信息
-        resp = self.session.get(url=self.api_url, params=self.general_query_data, headers=self.default_headers)
+        resp = self.session.get(url=self.api_url, params=general_query_data, headers=self.default_headers)
         json_data = json.loads(resp.content[len(callback_item) + 1: -1])
         # 响应示例
         # {"url":"","br":-1}
@@ -117,11 +124,12 @@ class GeneralMusicDownloader():
                 song_artist = songs[0]['artist'][0]  # 取第一位艺术家
                 pic_id = songs[0]['pic_id']
                 lyric_id = songs[0]['lyric_id']
+                album = songs[0]['album']
 
                 download_url = self.get_song_download_info_from_source(song_id=song_id, source=source)['url']
                 if download_url:
                     print 'find the download url from source %s' % source
-                    return download_url, song_name, song_artist, pic_id, lyric_id, source
+                    return download_url, song_name, song_artist, pic_id, lyric_id, source, album
                 else:
                     print 'Can not download from source %s' % source
                     continue  # 无法下载则切换源
@@ -146,15 +154,19 @@ class GeneralMusicDownloader():
 
     def get_and_download_pic(self, pic_id, song_name, song_artist, source):
         filename = '%s - %s.jpg' % (song_artist, song_name)
+        if os.path.exists(os.path.join(self.default_pic_folder, filename)):
+            print 'already download cover pic %s successfully' % filename
+            return True
         # 修改必要参数
+        general_query_data = self.general_query_data.copy()
         callback_item = 'jQuery%s_%s' % (random_int_string(21), timestamp_ms_string())
-        self.general_query_data['callback'] = callback_item
-        self.general_query_data['_'] = timestamp_ms_string()
-        self.general_query_data['id'] = pic_id
-        self.general_query_data['source'] = source
-        self.general_query_data['types'] = 'pic'
+        general_query_data['callback'] = callback_item
+        general_query_data['_'] = timestamp_ms_string()
+        general_query_data['id'] = pic_id
+        general_query_data['source'] = source
+        general_query_data['types'] = 'pic'
         # 获取下载信息
-        resp = self.session.get(url=self.api_url, params=self.general_query_data, headers=self.default_headers)
+        resp = self.session.get(url=self.api_url, params=general_query_data, headers=self.default_headers)
         json_data = json.loads(resp.content[len(callback_item) + 1: -1])
         if json_data['url']:
             urllib.urlretrieve(url=json_data['url'], filename=os.path.join(self.default_pic_folder, filename))
@@ -166,15 +178,19 @@ class GeneralMusicDownloader():
 
     def get_and_download_lyric(self, lyric_id, song_name, song_artist, source):
         filename = '%s - %s.lrc' % (song_artist, song_name)
+        if os.path.exists(os.path.join(self.default_download_folder, filename)):
+            print 'already download the lyric %s successfully' % filename
+            return True
         # 修改必要参数
+        general_query_data = self.general_query_data.copy()
         callback_item = 'jQuery%s_%s' % (random_int_string(21), timestamp_ms_string())
-        self.general_query_data['callback'] = callback_item
-        self.general_query_data['_'] = timestamp_ms_string()
-        self.general_query_data['id'] = lyric_id
-        self.general_query_data['source'] = source
-        self.general_query_data['types'] = 'lyric'
+        general_query_data['callback'] = callback_item
+        general_query_data['_'] = timestamp_ms_string()
+        general_query_data['id'] = lyric_id
+        general_query_data['source'] = source
+        general_query_data['types'] = 'lyric'
         # 获取下载信息
-        resp = self.session.get(url=self.api_url, params=self.general_query_data, headers=self.default_headers)
+        resp = self.session.get(url=self.api_url, params=general_query_data, headers=self.default_headers)
         json_data = json.loads(resp.content[len(callback_item) + 1: -1])
         if json_data['lyric']:
             with codecs.open(os.path.join(self.default_download_folder, filename), 'wb', 'utf-8') as f:
@@ -185,21 +201,47 @@ class GeneralMusicDownloader():
             print 'Can not download the lyric %s' % filename
             return False
 
-    def write_pic_to_song(self, song_name, song_artist):
+    def write_pic_to_song(self, song_name, song_artist, album):
         song_filename = os.path.join(self.default_download_folder, '%s - %s.mp3' % (song_artist, song_name))
         pic_filename = os.path.join(self.default_pic_folder, '%s - %s.jpg' % (song_artist, song_name))
 
-        audiofile = eyed3.load(song_filename)
-        if not audiofile.tag:
-            audiofile.initTag()
-        audiofile.tag.images.set(3, open(pic_filename, 'rb').read(), 'image/jpeg')
-        audiofile.tag.save()
+        try:
+            audiofile = eyed3.load(song_filename)
+            if not audiofile:
+                return
+            if not audiofile.tag:
+                audiofile.initTag()
+            audiofile.tag.images.set(3, open(pic_filename, 'rb').read(), 'image/jpeg')
+            audiofile.tag.artist = song_artist if not audiofile.tag.artist else audiofile.tag.artist
+            audiofile.tag.title = song_name if not audiofile.tag.title else audiofile.tag.title
+            audiofile.tag.album = album if not audiofile.tag.album else audiofile.tag.album
+            audiofile.tag.save()
+        except:
+            traceback.print_exc()
+            print song_name, song_artist, album, repr(song_name), repr(song_artist), repr(album)
+
+    def set_default_download_folder(self, path):
+        self.default_download_folder = path
+
+    def set_default_pic_folder(self, path):
+        self.default_pic_folder = path
+
+    def search_and_download_song(self, keyword, download_dir=None, pic_dir=None):
+        if download_dir:
+            self.set_default_download_folder(download_dir)
+        if pic_dir:
+            self.set_default_pic_folder(pic_dir)
+        download_url, song_name, song_artist, pic_id, lyric_id, source, album = self.get_song_download_url(keyword)
+        if self.download_song_from_url(download_url, song_name, song_artist):
+            self.get_and_download_lyric(lyric_id=lyric_id, song_name=song_name, song_artist=song_artist, source=source)
+            if self.get_and_download_pic(pic_id=pic_id, song_name=song_name, song_artist=song_artist, source=source):
+                self.write_pic_to_song(song_name, song_artist, album)
 
 
 if __name__ == '__main__':
     api = GeneralMusicDownloader()
-    download_url, song_name, song_artist, pic_id, lyric_id, source = api.get_song_download_url('幼稚完 林峰')
+    download_url, song_name, song_artist, pic_id, lyric_id, source, album = api.get_song_download_url('幼稚完 林峰')
     if api.download_song_from_url(download_url, song_name, song_artist):
         api.get_and_download_lyric(lyric_id=lyric_id, song_name=song_name, song_artist=song_artist, source=source)
         if api.get_and_download_pic(pic_id=pic_id, song_name=song_name, song_artist=song_artist, source=source):
-            api.write_pic_to_song(song_name, song_artist)
+            api.write_pic_to_song(song_name, song_artist, album)
